+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.9.3] - 25.10.2020
+
+* Added `toWidget()` extension method on `CommandResult`
+
+## [0.9.2] - 25.10.2020
+
+* Added `executeWithFuture` to use with `RefreshIndicator`
+* Added `toWidget()` method
+
 ## [0.9.1] - 24.08.2020
 
 * Shortened package description in pubspec